requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
flask-marshmallow==1.2.1
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
marshmallow==3.22.0
packaging==24.1
SQLAlchemy==2.0.35
typing_extensions==4.12.2
Werkzeug==3.0.4
circuitbreaker~=2.0.0
flask-swagger-ui~=4.11.1
Flask-Caching~=2.3.0
Flask-Limiter~=3.8.0
PyJWT~=2.9.0